Move upload control to header
