+++ app/albums/[id]/page.tsx
@@ -3,18 +3,16 @@
 /* eslint-disable @next/next/no-img-element */
 import Link from "next/link";
 import { useCallback, useState, useRef } from "react";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { getAlbum, addPhotosToAlbum } from "@/shared/api/albums";
-import { uploadPhotos } from "@/shared/api/photos";
+import { useQuery } from "@tanstack/react-query";
+import { getAlbum } from "@/shared/api/albums";
 import styles from "./album.module.css";
 
 type Props = { params: { id: string } };
 
 export default function AlbumView({ params }: Props) {
   const id = Number(params.id);
-  const queryClient = useQueryClient();
-  const [uploadProgress, setUploadProgress] = useState<number | null>(null);
   const [selectedIndex, setSelectedIndex] = useState<number | null>(null);
+  
   const {
     data: album,
     isLoading: albumLoading,
@@ -60,29 +58,6 @@ export default function AlbumView({ params }: Props) {
     touchStartX.current = null;
   };
 
-  const handlePhotoUpload = async (
-    e: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    const files = e.target.files;
-    if (!files?.length) return;
-    try {
-      setUploadProgress(0);
-      const uploaded = await uploadPhotos(
-        Array.from(files),
-        setUploadProgress,
-      );
-      if (uploaded.length) {
-        await addPhotosToAlbum(id, uploaded);
-        queryClient.invalidateQueries({ queryKey: ["album", id] });
-      }
-      e.target.value = "";
-    } catch (err) {
-      alert(err instanceof Error ? err.message : "Failed to upload photo");
-    } finally {
-      setUploadProgress(null);
-    }
-  };
-
   if (albumLoading) {
     return <p className={styles.status}>Loading album...</p>;
   }
@@ -93,13 +68,6 @@ export default function AlbumView({ params }: Props) {
   return (
     <main className={styles.main}>
       <h1 className={styles.title}>{album.title}</h1>
-      <input type="file" multiple onChange={handlePhotoUpload} />
-      {uploadProgress !== null && (
-        <div className={styles.uploadProgress}>
-          <progress value={uploadProgress} max={100} />
-          <span>{uploadProgress}%</span>
-        </div>
-      )}
       <div className={styles.photoGrid}>
         {photos.map(
           (photo, i) =>
