+++ app/home.module.css
@@ -162,13 +162,6 @@
   display: block;
 }
 
-.uploadProgress {
-  margin: 1rem 0;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
 .albumPlaceholder {
   width: 100%;
   height: 100%;
